+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - xScheduledTask:
   - Update existing Scheduled Task using SetScheduleTask
     instead of UnRegister/Register - See [Issue #134](https://github.com/PowerShell/xComputerManagement/issues/134).
+  - Fix error when creating Network Service Scheduled Task - See
+    [Issue #130](https://github.com/PowerShell/xComputerManagement/issues/130).
+  - Correct integration test configuration names for ExecuteAs tests.
 - Fix master branch AppVeyor badge link URL in README.MD - See [Issue #140](https://github.com/PowerShell/xComputerManagement/issues/140).
 - Fix deletion of scheduled task with unknown or empty task trigger.
   Get-TargetResource returns an empty ScheduleType string if the task